simple_a2a_agent/pyproject.toml
[project]
name = "simple-a2a-agent"
version = "0.1.0"
description = "Single A2A agent: weather (NWS MCP) + Airbnb search (MCP) in one server"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.2.0",
    "geopy>=2.4.1",
    "google-adk>=1.7.0",
    "httpx>=0.27.0",
    "litellm",
    "mcp>=1.5.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "uagents-core[a2a]==0.4.5a2",
]

[project.scripts]
simple-a2a-agent = "simple_a2a_agent.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
